Use toflar/fast-set
