+++ composer.json
@@ -27,9 +27,10 @@
     ],
     "require": {
         "php": "^8.1",
-        "ext-zlib": "*",
         "ext-intl": "*",
-        "ext-mbstring": "*"
+        "ext-mbstring": "*",
+        "ext-zlib": "*",
+        "toflar/fast-set": "dev-main"
     },
     "require-dev": {
         "phpstan/phpstan": "^2.0",
