+++ src/Tokenizer/LocaleConfiguration/AbstractPreconfiguredLocale.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Loupe\Matcher\Tokenizer\LocaleConfiguration;
 
-use Loupe\Matcher\Tokenizer\Decompounder\Dictionary\BinaryFileDictionary;
 use Loupe\Matcher\Tokenizer\Decompounder\Dictionary\DictionaryInterface;
+use Loupe\Matcher\Tokenizer\Decompounder\Dictionary\FastSetDictionary;
 use Loupe\Matcher\Tokenizer\Normalizer\Normalizer;
 use Loupe\Matcher\Tokenizer\Normalizer\NormalizerInterface;
 
@@ -15,7 +15,7 @@ abstract class AbstractPreconfiguredLocale implements LocaleConfigurationInterfa
 
     public function __construct()
     {
-        $this->dictionary = BinaryFileDictionary::create(
+        $this->dictionary = FastSetDictionary::create(
             $this->getLocale(),
             __DIR__ . '/../../../dictionaries/' . $this->getLocale()->toString()
         );
